# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SOLOS+ ESPORTZ API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List events
      responses:
        "200":
          description: List of events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"

    post:
      operationId: createEvent
      tags: [events]
      summary: Create event (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
  /events/{id}:
    get:
      operationId: getEvent
      tags: [events]
      summary: Get event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"

    patch:
      operationId: updateEvent
      tags: [events]
      summary: Update event (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"

    delete:
      operationId: deleteEvent
      tags: [events]
      summary: Delete event (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # SCRIMS
  /scrims:
    get:
      operationId: listScrims
      tags: [scrims]
      summary: List scrims
      responses:
        "200":
          description: List of scrims
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Scrim"

    post:
      operationId: createScrim
      tags: [scrims]
      summary: Create scrim (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScrimInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scrim"
  /scrims/{id}:
    get:
      operationId: getScrim
      tags: [scrims]
      summary: Get scrim details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Scrim
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scrim"

    patch:
      operationId: updateScrim
      tags: [scrims]
      summary: Update scrim (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScrimUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scrim"

    delete:
      operationId: deleteScrim
      tags: [scrims]
      summary: Delete scrim (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /scrims/{id}/register:
    post:
      operationId: registerForScrim
      tags: [scrims]
      summary: Register for a scrim
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

    delete:
      operationId: unregisterFromScrim
      tags: [scrims]
      summary: Unregister from a scrim
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Unregistered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /scrims/{id}/result:
    patch:
      operationId: postScrimResult
      tags: [scrims]
      summary: Post scrim result (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScrimResultInput"
      responses:
        "200":
          description: Result posted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scrim"

  # FEED
  /feed:
    get:
      operationId: listFeedPosts
      tags: [feed]
      summary: List feed posts
      responses:
        "200":
          description: Feed
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FeedPost"

    post:
      operationId: createFeedPost
      tags: [feed]
      summary: Create feed post (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeedPostInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedPost"
  /feed/{id}:
    delete:
      operationId: deleteFeedPost
      tags: [feed]
      summary: Delete feed post (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /feed/{id}/like:
    post:
      operationId: likeFeedPost
      tags: [feed]
      summary: Like a feed post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Liked/unliked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LikeResponse"

  # LEADERBOARD
...
components:
  schemas:
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string

    # AUTH
...
    Event:
      type: object
      required: [id, title, description, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        imageUrl:
          type: ["string", "null"]
        link:
          type: ["string", "null"]
        eventDate:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    EventInput:
      type: object
      required: [title, description]
      properties:
        title:
          type: string
        description:
          type: string
        imageUrl:
          type: ["string", "null"]
        link:
          type: ["string", "null"]
        eventDate:
          type: ["string", "null"]
    EventUpdate:
      type: object
      properties:
        title:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        link:
          type: ["string", "null"]
        eventDate:
          type: ["string", "null"]

    # SCRIMS
    Scrim:
      type: object
      required: [id, title, status, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        link:
          type: ["string", "null"]
        scheduledAt:
          type: ["string", "null"]
        status:
          type: string
          enum: [UPCOMING, ONGOING, COMPLETED, CANCELLED]
        result:
          type: ["string", "null"]
        participants:
          type: array
          items:
            $ref: "#/components/schemas/ScrimParticipant"
        createdAt:
          type: string
    ScrimParticipant:
      type: object
      required: [userId, username, clanTag]
      properties:
        userId:
          type: integer
        username:
          type: string
        clanTag:
          type: string
        avatarUrl:
          type: ["string", "null"]
    ScrimInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        link:
          type: ["string", "null"]
        scheduledAt:
          type: ["string", "null"]
    ScrimUpdate:
      type: object
      properties:
        title:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        link:
          type: ["string", "null"]
        scheduledAt:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
    ScrimResultInput:
      type: object
      required: [result]
      properties:
        result:
          type: string
        status:
          type: string
          enum: [COMPLETED, CANCELLED]

    # FEED
    FeedPost:
      type: object
      required: [id, content, type, likeCount, isLikedByMe, createdAt]
      properties:
        id:
          type: integer
        content:
          type: string
        type:
          type: string
          enum: [ANNOUNCEMENT, EVENT, SCRIM_RESULT, IMAGE, NEWS]
        imageUrl:
          type: ["string", "null"]
        link:
          type: ["string", "null"]
        authorId:
          type: ["integer", "null"]
        authorName:
          type: ["string", "null"]
        authorAvatar:
          type: ["string", "null"]
        likeCount:
          type: integer
        isLikedByMe:
          type: boolean
        createdAt:
          type: string
    FeedPostInput:
      type: object
      required: [content, type]
      properties:
        content:
          type: string
        type:
          type: string
          enum: [ANNOUNCEMENT, EVENT, SCRIM_RESULT, IMAGE, NEWS]
        imageUrl:
          type: ["string", "null"]
        link:
          type: ["string", "null"]
    LikeResponse:
      type: object
      required: [liked, likeCount]
      properties:
        liked:
          type: boolean
        likeCount:
          type: integer

    # LEADERBOARD
...