# lib/api-zod/src/generated/types/likeResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SOLOS+ ESPORTZ API
 * OpenAPI spec version: 0.1.0
 */

export interface LikeResponse {
  liked: boolean;
  likeCount: number;
}