# lib/api-zod/src/generated/types/scrimInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SOLOS+ ESPORTZ API
 * OpenAPI spec version: 0.1.0
 */

export interface ScrimInput {
  title: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  link?: string | null;
  /** @nullable */
  scheduledAt?: string | null;
}